# Sort position links for faster non-equi joins