--- a/presto-main/src/main/java/com/facebook/presto/SystemSessionProperties.java
+++ b/presto-main/src/main/java/com/facebook/presto/SystemSessionProperties.java
@@ -62,6 +62,7 @@ public final class SystemSessionProperties
     public static final String INITIAL_SPLITS_PER_NODE = "initial_splits_per_node";
     public static final String SPLIT_CONCURRENCY_ADJUSTMENT_INTERVAL = "split_concurrency_adjustment_interval";
     public static final String OPTIMIZE_METADATA_QUERIES = "optimize_metadata_queries";
+    public static final String FAST_INEQUALITY_JOIN = "fast_inequality_join";
     public static final String QUERY_PRIORITY = "query_priority";
     public static final String SPILL_ENABLED = "spill_enabled";
     public static final String OPERATOR_MEMORY_LIMIT_BEFORE_SPILL = "operator_memory_limit_before_spill";
@@ -242,6 +243,11 @@ public final class SystemSessionProperties
                         featuresConfig.isJoinReorderingEnabled(),
                         false),
                 booleanSessionProperty(
+                        FAST_INEQUALITY_JOIN,
+                        "Experimental: Use faster handling of inequality join if it is possible",
+                        featuresConfig.isFastInequalityJoins(),
+                        false),
+                booleanSessionProperty(
                         COLOCATED_JOIN,
                         "Experimental: Use a colocated join when possible",
                         featuresConfig.isColocatedJoinsEnabled(),
@@ -396,6 +402,11 @@ public final class SystemSessionProperties
     public static boolean planWithTableNodePartitioning(Session session)
     {
         return session.getSystemProperty(PLAN_WITH_TABLE_NODE_PARTITIONING, Boolean.class);
+    }
+
+    public static boolean isFastInequalityJoin(Session session)
+    {
+        return session.getSystemProperty(FAST_INEQUALITY_JOIN, Boolean.class);
     }
 
     public static boolean isJoinReorderingEnabled(Session session)