--- a/presto-main/src/main/java/com/facebook/presto/operator/JoinFilterFunction.java
+++ b/presto-main/src/main/java/com/facebook/presto/operator/JoinFilterFunction.java
@@ -15,7 +15,11 @@ package com.facebook.presto.operator;
 
 import com.facebook.presto.spi.Page;
 
+import java.util.Optional;
+
 public interface JoinFilterFunction
 {
     boolean filter(int leftAddress, int rightPosition, Page rightPage);
+
+    Optional<Integer> getSortChannel();
 }