--- a/presto-main/src/main/java/com/facebook/presto/operator/PagesHashStrategy.java
+++ b/presto-main/src/main/java/com/facebook/presto/operator/PagesHashStrategy.java
@@ -92,4 +92,6 @@ public interface PagesHashStrategy
      * Checks if any of the hashed columns is null
      */
     boolean isPositionNull(int blockIndex, int blockPosition);
+
+    int compare(int leftBlockIndex, int leftBlockPosition, int rightBlockIndex, int rightBlockPosition);
 }