--- a/presto-main/src/main/java/com/facebook/presto/operator/StandardJoinFilterFunction.java
+++ b/presto-main/src/main/java/com/facebook/presto/operator/StandardJoinFilterFunction.java
@@ -19,6 +19,7 @@ import com.google.common.collect.ImmutableList;
 import it.unimi.dsi.fastutil.longs.LongArrayList;
 
 import java.util.List;
+import java.util.Optional;
 
 import static com.facebook.presto.operator.SyntheticAddress.decodePosition;
 import static com.facebook.presto.operator.SyntheticAddress.decodeSliceIndex;
@@ -32,11 +33,13 @@ public class StandardJoinFilterFunction
     private final InternalJoinFilterFunction filterFunction;
     private final LongArrayList addresses;
     private final List<Block[]> pages;
+    private final Optional<Integer> sortChannel;
 
-    public StandardJoinFilterFunction(InternalJoinFilterFunction filterFunction, LongArrayList addresses, List<List<Block>> channels)
+    public StandardJoinFilterFunction(InternalJoinFilterFunction filterFunction, LongArrayList addresses, List<List<Block>> channels, Optional<Integer> sortChannel)
     {
         this.filterFunction = requireNonNull(filterFunction, "filterFunction can not be null");
         this.addresses = requireNonNull(addresses, "addresses is null");
+        this.sortChannel = requireNonNull(sortChannel, "sortChannel is null");
 
         requireNonNull(channels, "channels can not be null");
         ImmutableList.Builder<Block[]> pagesBuilder = ImmutableList.builder();
@@ -61,6 +64,12 @@ public class StandardJoinFilterFunction
         int blockPosition = decodePosition(pageAddress);
 
         return filterFunction.filter(blockPosition, getLeftBlocks(blockIndex), rightPosition, rightPage.getBlocks());
+    }
+
+    @Override
+    public Optional<Integer> getSortChannel()
+    {
+        return sortChannel;
     }
 
     private Block[] getLeftBlocks(int leftBlockIndex)