… presto-benchmark/src/test/java/com/facebook/presto/benchmark/BenchmarkDecimalAggregation.java
… presto-benchmark/src/test/java/com/facebook/presto/benchmark/BenchmarkInequalityJoin.java
… presto-benchmark/src/test/java/com/facebook/presto/benchmark/MemoryLocalQueryRunner.java
… presto-cassandra/src/test/java/com/facebook/presto/cassandra/TestCassandraDistributed.java
… presto-main/src/main/java/com/facebook/presto/operator/SortedPositionLinks.java
… presto-main/src/main/java/com/facebook/presto/sql/gen/JoinFilterFunctionCompiler.java
… presto-main/src/main/java/com/facebook/presto/sql/planner/LocalExecutionPlanner.java
… presto-main/src/main/java/com/facebook/presto/sql/planner/SortExpressionExtractor.java
… presto-main/src/test/java/com/facebook/presto/operator/BenchmarkHashBuildAndJoinOperators.java
… presto-main/src/test/java/com/facebook/presto/operator/TestHashJoinOperator.java
… presto-main/src/test/java/com/facebook/presto/operator/TestPositionLinks.java
… presto-main/src/test/java/com/facebook/presto/sql/analyzer/TestFeaturesConfig.java
… presto-main/src/test/java/com/facebook/presto/sql/gen/TestJoinCompiler.java
… presto-main/src/test/java/com/facebook/presto/sql/gen/TestJoinProbeCompiler.java
… presto-main/src/test/java/com/facebook/presto/sql/planner/TestSortExpressionExtractor.java
… presto-tests/src/main/java/com/facebook/presto/tests/AbstractTestQueries.java