This is a implementation of:

https://github.com/prestodb/presto/issues/6922

From preliminary benchmarks this change didn't introduce regression in equality joins and for some in-equality joins this gives orders of magnitude speed ups.

`BenchmarkInequalityJoin` speeds up from ~210ms down to ~50ms and existing `BenchmarkHashBuildAndJoinOperators` do not show a visible performance regression:

```
AFTER

Benchmark                                              (hashColumns)  (matchRate)  Mode  Cnt    Score    Error  Units
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash        varchar          N/A  avgt   10   56.792 ±  7.107  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash         bigint          N/A  avgt   10   35.530 ±  2.304  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash            all          N/A  avgt   10   69.249 ±  7.824  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash         varchar          0.1  avgt   10  100.534 ±  8.280  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash         varchar            1  avgt   10  127.782 ±  5.132  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash         varchar            2  avgt   10  118.533 ± 13.199  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash          bigint          0.1  avgt   10   79.058 ±  4.773  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash          bigint            1  avgt   10  100.670 ± 10.200  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash          bigint            2  avgt   10   74.568 ±  2.282  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash             all          0.1  avgt   10  119.906 ±  1.313  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash             all            1  avgt   10  156.546 ± 12.703  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash             all            2  avgt   10  138.223 ±  7.201  ms/op

BEFORE:

Benchmark                                              (hashColumns)  (matchRate)  Mode  Cnt    Score    Error  Units
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash        varchar          N/A  avgt   10   54.626 ±  5.888  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash         bigint          N/A  avgt   10   35.253 ±  1.619  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash            all          N/A  avgt   10   73.002 ±  8.299  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash         varchar          0.1  avgt   10   93.097 ±  4.915  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash         varchar            1  avgt   10  130.529 ±  7.421  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash         varchar            2  avgt   10  114.086 ± 10.210  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash          bigint          0.1  avgt   10   71.442 ±  4.633  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash          bigint            1  avgt   10   92.942 ± 12.446  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash          bigint            2  avgt   10   77.394 ±  2.632  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash             all          0.1  avgt   10  111.427 ±  3.545  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash             all            1  avgt   10  150.452 ± 15.728  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkJoinHash             all            2  avgt   10  130.621 ±  3.306  ms/op

```

There are couple of TODOs (future work?):

- `compare` method may not fit to `PagesHashStrategy`
- improve decision between choosing `SortedPositionLinks` over `PositionLinksList`
- maybe add `NoopPositionLinks` (that returns always `-1`) if all rows in build table have unique key (this would reduce memory usage and it might slightly improve performance)
- add similar logic to cross join