@pnowojski 2017-03-27
@dain I have fixed concurrency issues without refactor (as discussed) and it's read for review.