@losipiuk 2017-03-27
@dain I will give it one more read but feel free to go through it at the same time.