@dain 2017-03-31
Review status:
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/353394099c6af4217885e8d281db7a08fbbee33e Collect output of the memory benchmarks
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/4486d5e1f2d3e25965cfbe03d3a20dbddd3da688 Use SortedPositionLinks in LookupJoinOperator
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/f20ac3adef08c5e9dc8559647ca4cb015cbeb008 Add compare method to PagesHashStrategy
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/29b2f670c0f052c57151b5985aaeaa0fee9517cd Extract sort channel from filter function
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/87b119a160025ea8cfa97a00f515b2c6b7c104f4 Add tests queries for sorting position links
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/3f9d0df02da533a51303095c84fa5e7224185096 Add unit tests for SortExpressionExtractor
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/2ee71114d34ab4d0e6d533366e73d47001e78102 Add SortExpressionExtractor
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/3d33e0bccf928261b95117ec76b80347d2366f6a Add unit tests for PositionLinks
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/0a53e60b459f9ea97e77316600352358d8c9ab66 Update presto-benchto-benchmarks README file
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/5027535b98e837fb2016e7c3656594839b9cc9c0 Make schema names categorial templates
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/d4f25a26a10624978237f08d1924b2aeaffc64ec Update benchto-driver dependency to 0.3
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/b2e85b691ef841ae5757710d95bf8a1e7e9ae6f8 Implemented SortedPositionLinks with binary search
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/ca87b79925a8656be13453cced002c4634944575 Implement ArrayPositionLinks
- [ ]  https://github.com/prestodb/presto/pull/7097/commits/6b538a7fa3d6af61c0815ca47c8c676f1bf8d23b Add PositionLinks interface