@pnowojski 2017-04-14
Applied comments.

I have added some more benchmark cases to see if there is a performance drop for very short position links (larger number of buckets) and surprising I haven't seen that. So indeed I think we can turn this on by default.

Here are the results. Number of elements in build table is `60175`, so position links length is equal to `60175 / buckets`. Of course the larger position links chain, the better case for the binary search.
```
Benchmark                              (buckets)  (fastInequalityJoin)  (filterOutCoefficient)  Mode  Cnt     Score    Error  Units
BenchmarkInequalityJoin.benchmarkJoin        100                   true                      10  avgt   30    79.657 ±  4.545  ms/op
BenchmarkInequalityJoin.benchmarkJoin        100                  false                      10  avgt   30  1367.266 ± 28.193  ms/op
BenchmarkInequalityJoin.benchmarkJoin       1000                   true                      10  avgt   30    70.507 ±  4.654  ms/op
BenchmarkInequalityJoin.benchmarkJoin       1000                  false                      10  avgt   30   204.847 ±  6.114  ms/op
BenchmarkInequalityJoin.benchmarkJoin      10000                   true                      10  avgt   30    65.660 ±  4.196  ms/op
BenchmarkInequalityJoin.benchmarkJoin      10000                  false                      10  avgt   30    82.278 ±  3.722  ms/op
BenchmarkInequalityJoin.benchmarkJoin      60000                   true                      10  avgt   30    61.385 ±  3.970  ms/op
BenchmarkInequalityJoin.benchmarkJoin      60000                  false                      10  avgt   30    65.466 ±  5.425  ms/op
```