@dain 2017-04-14
The "performance drop" you are seeing is in the margin of error, so it doesn't mean anything.