@pnowojski 2017-04-14
Maybe I wasn't clear enough. I meant, that I expected to see some performance regression in some edge cases, but it wasn't visible in benchmarks (which is good :) )